:ok_hand: :truck: Move `App.js` and `Root.js` to `app` directory
