+++ app/Root.js
@@ -2,7 +2,8 @@
 import React, { Component } from 'react';
 import { Provider } from 'react-redux';
 import { ConnectedRouter } from 'react-router-redux';
-import Routes from '../routes';
+
+import Routes from './routes';
 
 type Props = {
   store: {},
