+++ app/routes.js
@@ -1,7 +1,8 @@
 /* eslint flowtype-errors/show-errors: 0 */
 import React from 'react';
 import { Switch, Route } from 'react-router';
-import App from './screens/App';
+
+import App from './App';
 import Login from './screens/Login';
 
 export default () => (
